+++ lib/Progress/VariantSummaryFormatter.php
@@ -17,7 +17,7 @@ EOT
     public const BASELINE_FORMAT = <<<'EOT'
 "[" ~ 
 "Mo" ~ mode(variant.time.avg) as time ~
-" <fg=magenta;bg=black>vs</> " ~ 
+" vs " ~ 
 "Mo" ~ mode(baseline.time.avg) as time ~ "] " ~ 
 percent_diff(mode(baseline.time.avg), mode(variant.time.avg), (rstdev(variant.time.avg) * 2)) ~
 " (±" ~ rstdev(variant.time.avg) ~ "%)"